@@ -1,5 +1,5 @@
 // Hex JavaScript decoder
-// Copyright (c) 2008-2023 Lapo Luchini <lapo@lapo.it>
+// Copyright (c) 2008-2024 Lapo Luchini <lapo@lapo.it>
 
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -13,18 +13,12 @@
 // ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
 // OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 
-(typeof define != 'undefined' ? define : function (factory) { 'use strict';
-    if (typeof module == 'object') module.exports = factory();
-    else window.hex = factory();
-})(function () {
-'use strict';
-
 const
     haveU8 = (typeof Uint8Array == 'function');
 
 let decoder; // populated on first usage
 
-class Hex {
+export class Hex {
 
     /**
      * Decodes an hexadecimal value.
@@ -73,7 +67,3 @@ class Hex {
     }
 
 }
-
-return Hex;
-
-});
